more changes

+++ web/app/controllers/cycles_controller.rb
@@ -3,7 +3,7 @@ class CyclesController < ApplicationController
   
   def create
     @cycle = Cycle.new(cycle_params)
-    @cycle.protocol_id = params[:protocol_id]
+    @cycle.experiment_id = params[:experiment_id]
     @cycle.next_component_id = params[:next_component]
     @step = Step.new(step_params)
     @cycle.children << @step
@@ -37,9 +37,15 @@ class CyclesController < ApplicationController
   
   def destroy
     @cycle = Cycle.find(params[:id])
-    ret = @cycle.destroy
-    respond_to do |format|
-       format.json { render :json => @cycle,:status => (ret)? :ok : :unprocessable_entity}
+    if @cycle.experiment.components.count > 1
+      ret = @cycle.destroy
+      respond_to do |format|
+        format.json { render :json => @cycle,:status => (ret)? :ok : :unprocessable_entity}
+      end
+    else
+      respond_to do |format|
+        format.json { render :json => {:base => "lastnode"},:status => :unprocessable_entity}
+      end
     end
   end
